feat: add get reactive rest client config
